requirements.txt
Pillow
requests
torch
numpy
pycocotools
coco-types
matplotlib
#git+https://github.com/facebookresearch/detectron2.git
ftfy
regex
tqdm
torchvision
git+https://github.com/openai/CLIP.git